+++ CHANGELOG.md
@@ -10,6 +10,12 @@
 - [Rust](https://github.com/moonrepo/rust-plugin/blob/master/CHANGELOG.md)
 - [TOML schema](https://github.com/moonrepo/schema-plugin/blob/master/CHANGELOG.md)
 
+## Unreleased
+
+#### ⚙️ Internal
+
+- Improved command execution. May see some slight performance gains.
+
 ## 0.36.2
 
 #### 🚀 Updates