+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 #### ⚙️ Internal
 
 - Improved command execution. May see some slight performance gains.
-- Upgraded wasmtime to v21 (from v17).
+- Updated wasmtime to v21 (from v17).
+- Updated Rust to v1.78.
 
 ## 0.36.2